requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
tensorflow-cpu==2.20.0
numpy==2.1.3
opencv-python==4.12.0.88
pillow==11.3.0
pydantic==2.12.4
pydantic_core==2.41.5
python-multipart==0.0.20
starlette==0.49.3
typing_extensions==4.14.1
gdown==4.7.1
tensorflow==2.15.0
keras==2.15.0



